DAOS-4269 bio: bump max ops per io channel to 4k
The max blob read/write ops per io channel is 512 by default, it's
not enough for aggregation: Aggregation merge window size is 8MB,
let's assume an extreme case that the merge window consists of 4k
adjacent NVMe extents, then there will be 2k blob read ops on
window flush.

There could be more blob reads on window flush when the window
consists of overlapped NVMe extents, so we have to limit the number
of NVMe extents in a merge window.

This patch bumped the max ops to 4k in BIO, and limited the NVMe
extent number in merge window to 2k. (leaving margin for IO requests)

Signed-off-by: Niu Yawei <yawei.niu@intel.com>